requirements.txt
Cython==0.29.2
eventlet==0.24.1
Flask-SocketIO==3.1.2
imutils==0.5.2
motmetrics==1.1.3
mysql-connector-python==8.0.12
numba==0.42.0
numpy==1.16.0
opencv-python==4.0.0.21
opencv-contrib-python==4.0.0.21
pairing==0.1.3
pandas==0.23.4
Pillow==5.4.1
redis==3.0.1
scipy==1.2.0
untangle==1.1.1
easydict==1.9
torch==1.0.1
torchvision==0.2.1
tqdm==4.30.0
beautifulsoup4==4.7.1
requests==2.21.0
urllib5==5.0.0
